requirements.txt
beautifulsoup4==4.9.1
bs4==0.0.1
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
cryptography==2.9.2
decorator==4.4.2
dnslib==0.9.13
dnspython==1.16.0
future==0.18.2
idna==2.9
pycparser==2.20
pymongo==3.10.1
python-telegram-bot==12.7
PythonDNS==0.1
requests==2.23.0
six==1.15.0
soupsieve==2.0.1
tornado==6.0.4
urllib3==1.25.9
